# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Transaction:
    properties:
      _id:
        type: string
      category:
        type: integer
      currency:
        type: string
      description:
        type: string
      method:
        $ref: '#/definitions/models.TransactionMethod'
      price:
        type: number
      title:
        type: string
      transaction_date:
        type: string
      user_id:
        type: string
    type: object
  models.TransactionMethod:
    properties:
      method_id:
        type: string
      type:
        type: integer
    type: object
...
  requests.ID:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  requests.TransactionCreate:
    properties:
      category:
        type: integer
      currency:
        type: string
      description:
        type: string
      method:
        $ref: '#/definitions/requests.TransactionMethod'
      price:
        type: number
      title:
        type: string
      transaction_date:
        type: string
    required:
    - category
    - currency
    - price
    - title
    - transaction_date
    type: object
  requests.TransactionMethod:
    properties:
      method_id:
        type: string
      type:
        type: integer
    required:
    - method_id
    - type
    type: object
  requests.TransactionUpdate:
    properties:
      category:
        type: integer
      currency:
        type: string
      delete_method:
        type: boolean
      description:
        type: string
      id:
        type: string
      method:
        $ref: '#/definitions/requests.TransactionMethod'
      price:
        type: number
      title:
        type: string
      transaction_date:
        type: string
    required:
    - id
    type: object
...
  responses.TransactionCategoryStat:
    properties:
      _id:
        type: integer
      total_transaction:
        type: number
    type: object
  responses.TransactionCategoryStats:
    properties:
      category_list:
        items:
          $ref: '#/definitions/responses.TransactionCategoryStat'
        type: array
      currency:
        type: string
      total_transaction:
        type: number
    type: object
  responses.TransactionDailyStats:
    properties:
      currency:
        type: string
      date:
        type: string
      total_transaction:
        type: number
    type: object
  responses.TransactionStats:
    properties:
      category_stats:
        $ref: '#/definitions/responses.TransactionCategoryStats'
      daily_stats:
        items:
          $ref: '#/definitions/responses.TransactionDailyStats'
        type: array
      total_expense:
        type: number
      total_income:
        type: number
    type: object
  responses.TransactionTotal:
    properties:
      currency:
        type: string
      total_transaction:
        type: number
    type: object
...
host: rocky-reaches-65250.herokuapp.com
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Kantan.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://rocky-reaches-65250.herokuapp.com/terms/
  title: Kantan Investment Manager API
  version: "1.0"
paths:
...
  /transaction:
    delete:
      consumes:
      - application/json
      description: Deletes transaction by id
      parameters:
      - description: ID
        in: body
        name: ID
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete transaction by transaction id
      tags:
      - transaction
    get:
      consumes:
      - application/json
      description: Returns transactions by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Transactions by User ID
      tags:
      - transaction
    post:
      consumes:
      - application/json
      description: Creates transaction
      parameters:
      - description: Transaction Create
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/requests.TransactionCreate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Transaction
      tags:
      - transaction
    put:
      consumes:
      - application/json
      description: Updates transaction
      parameters:
      - description: Transaction Update
        in: body
        name: transactionupdate
        required: true
        schema:
          $ref: '#/definitions/requests.TransactionUpdate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: Unauthorized update
          schema:
            type: string
        "404":
          description: Couldn't find user
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Transaction
      tags:
      - transaction
  /transaction/all:
    delete:
      consumes:
      - application/json
      description: Deletes all transaction by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete all transaction by user id
      tags:
      - transaction
  /transaction/stats:
    get:
      consumes:
      - application/json
      description: Returns transaction statistics by interval
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Transaction Statistics by Interval
      tags:
      - transaction
  /transaction/total:
    get:
      consumes:
      - application/json
      description: Returns total transaction value by interval
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TransactionTotal'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Total Transaction Value by Interval
      tags:
      - transaction
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"